Extract 'prepare' code to function
So that we can call it from other places, not just the CLI
